HMA-6127:new component (EditableListView) to library
