+++ sample/src/main/java/uk/gov/hmrc/components/sample/organisms/EditableListViewModel.kt
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2023 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package uk.gov.hmrc.components.sample.organisms
+
+import uk.gov.hmrc.components.organism.editable.EditableListView
+
+data class EditableListViewModel(
+    override var name: String,
+    override var value: String,
+    override var buttonText: String,
+    override var valueContentDescription: String,
+    override val onClickListener: (Int) -> Unit
+) : EditableListView.EditableItem
